Revert "Distinguished two types of tokens. Typographic tokens that provide column and length with the line index, and normal ones which do not."
This reverts commit e091bed44a98e4a5702edae2ee0ae0fe35ac26ad.
