+++ jplag.frontend.java/src/main/java/de/jplag/java/JavaToken.java
@@ -1,8 +1,8 @@
 package de.jplag.java;
 
-import de.jplag.TypographicToken;
+import de.jplag.Token;
 
-public class JavaToken extends TypographicToken implements JavaTokenConstants {
+public class JavaToken extends Token implements JavaTokenConstants {
 
     public JavaToken(int type, String file, int line, int col, int length) {
         super(type, file, line, col, length);
