+++ jplag.frontend.text/src/main/java/de/jplag/text/TextToken.java
@@ -1,8 +1,8 @@
 package de.jplag.text;
 
-import de.jplag.TypographicToken;
+import de.jplag.Token;
 
-public class TextToken extends TypographicToken {
+public class TextToken extends Token {
     private static final long serialVersionUID = 4301179216570538972L;
 
     private static int getSerial(String text, Parser parser) {
